Create TryToParseDouble.java
The Double.parseDouble() method requires a String argument, but
it fails if the String cannot be converted to a floating-point number.
Write an application in which you try accepting a double input from
a user and catch a NumberFormatException if one is thrown. The
catch block forces the number to 0 and displays an appropriate error
message. Following the catch block, display the number